Feat : Optimize collision detect, Ignoring collisions between agents
이름 변경 navTrace.cs -> SpecialMonsterAI.cs
SP1 AI추적 동작 SpecialMonster1.cs에서 호출 방식으로 변경
일반 몬스터 중력 변경시 충돌 감지 최적화
-> 오작동 할 경우 풀링 반환
일반 몬스터, SP1 끼리 충돌 무시
