+++ Assets/UserFolder/Prefab/Monster/NormalZombie/UrbanZombie.prefab
@@ -6671,7 +6671,7 @@ NavMeshAgent:
   m_Height: 1.75
   m_BaseOffset: 0
   m_WalkableMask: 29
-  m_ObstacleAvoidanceType: 4
+  m_ObstacleAvoidanceType: 2
 --- !u!114 &-8942009122545042035
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -6700,8 +6700,8 @@ MonoBehaviour:
   climbingDetectLayer:
     serializedVersion: 2
     m_Bits: 417792
-  castHeight: 2
-  castRadius: 0.5
+  castHeight: 2.35
+  castRadius: 0.75
 --- !u!114 &6681876166821450025
 MonoBehaviour:
   m_ObjectHideFlags: 0
