+++ Assets/UserFolder/Script/Monster/SpecialMonster/SpecialMonster1/SpecialMonsterAI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f2336000a1c92d34ba122944f27e6636
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
