+++ ProjectSettings/DynamicsManager.asset
@@ -18,7 +18,7 @@ PhysicsManager:
   m_ClothInterCollisionDistance: 0.1
   m_ClothInterCollisionStiffness: 0.2
   m_ContactsGeneration: 1
-  m_LayerCollisionMatrix: cfffffffcfffffffcfffffffcfffffff4010e8ff4010e8ffffffffffcfffffffcfffffffcfffffffcfffffffcfffffffffffffffcfffffffcfffffffcfffffffcfffffffcfffffffcfffffffffffffffcfffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff
+  m_LayerCollisionMatrix: cbffffffcbffffff4810e8ffcfffffff4010e8ff4010e8ffffffffffcbffffffcbfbffffcbffffffcbfeffffcbffffffffffffffcbffffffcbffffffcbffffffcbffffffcbffffffcbffffffffffffffcbffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff
   m_AutoSimulation: 1
   m_AutoSyncTransforms: 0
   m_ReuseCollisionCallbacks: 0
